retirada de código inútil, preparação inicial para aplicar banco na tela de revisão do artigo
