+++ app/src/main/java/com/example/cedri_app/ApprovalsActivity.kt
@@ -46,8 +46,6 @@ class ApprovalsActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_approvals)
 
-        //val token = NetworkUtils.getToken(getIntent().getExtras())
-
         val CursorDatabase = myDB.getTokenFromDatabase()
         var token : String = ""
 
@@ -58,17 +56,12 @@ class ApprovalsActivity : AppCompatActivity() {
         backImageButtonArticlesReview.setOnClickListener {
             val intent = Intent(this, MenuActivity::class.java)
             startActivity(intent)
-            //intent.putExtra("token", token)
             finish()
         }
-
-        // retriveBackendData()
 
         layoutManager = LinearLayoutManager(this)
         approval_recyclerView.layoutManager = layoutManager
         getPage(token)
-
-        Log.e("E", "--->" + "FUDEU")
 
         approval_recyclerView.addOnScrollListener(object : RecyclerView.OnScrollListener() {
             override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
@@ -78,15 +71,11 @@ class ApprovalsActivity : AppCompatActivity() {
                 val total = adapter.itemCount
 
                 if (!isLoading) {
-
                     if ((visibleItemCount + pastVisibleItem) >= total) {
                         page++
-                        // BUSCAR NO BACKEND DADOS DA PROXIMA PAGINA
                         getPage(token)
                     }
-
                 }
-
                 super.onScrolled(recyclerView, dx, dy)
             }
         })
@@ -99,10 +88,7 @@ class ApprovalsActivity : AppCompatActivity() {
         val end = articlesApprovalsList.size + limit
 
         retriveBackendData(token)
-        /*
-        for (i in start until end) {
-            articlesApprovalsList.add(articlesApprovalsList[i])
-        }*/
+
         Handler().postDelayed({
             if (::adapter.isInitialized) {
                 adapter.notifyDataSetChanged()
@@ -115,17 +101,13 @@ class ApprovalsActivity : AppCompatActivity() {
 
                     val intent = Intent(this, ArticleReviewActivity::class.java)
                     startActivity(intent)
-                    //intent.putExtra("token", NetworkUtils.getToken(getIntent().getExtras()))
                     finish()
-
-
                 }
                 approval_recyclerView.adapter = adapter
             }
             isLoading = false
             approval_progressBar.visibility = View.GONE
         }, 2000)
-
     }
 
     fun retriveBackendData(token : String){
@@ -146,7 +128,6 @@ class ApprovalsActivity : AppCompatActivity() {
                 val projects = body?.getData()?.elements
                 val size = projects?.size
 
-
                 if (projects != null){
                     // executa se o limite de paginas n ultrapassar o enviado pelo backend
                     if (body.getData().pagesTotal >= page){
@@ -156,11 +137,8 @@ class ApprovalsActivity : AppCompatActivity() {
                     } else {
                         return
                     }
-
                 }
-
             }
-
         })
     }
