+++ app/src/main/java/com/example/cedri_app/ArticleReviewActivity.kt
@@ -3,15 +3,23 @@ package com.example.cedri_app
 import android.content.Intent
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import com.example.cedri_app.database.DatabaseHandler
 import kotlinx.android.synthetic.main.activity_article_review.*
 
 class ArticleReviewActivity : AppCompatActivity() {
+
+    var myDB = DatabaseHandler(this)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_article_review)
 
-        val token = NetworkUtils.getToken( getIntent().getExtras() )
+        val CursorDatabase = myDB.getTokenFromDatabase()
+        var token : String = ""
+
+        if(CursorDatabase.moveToFirst()){
+            token = CursorDatabase.getString(CursorDatabase.getColumnIndex("token"))
+        }
 
         backImageButtonSingleArticleReview.setOnClickListener {
             val intent = Intent(this, ApprovalsActivity::class.java)
@@ -21,7 +29,6 @@ class ArticleReviewActivity : AppCompatActivity() {
 
         buttonPerfilAutorTelaResumoArtigo.setOnClickListener {
             val intent = Intent(this, AuthorPerfilActivity::class.java)
-            intent.putExtra("token", token)
             startActivity(intent)
             finish()
         }
