+++ app/src/main/java/com/example/cedri_app/MenuActivity.kt
@@ -12,46 +12,30 @@ import kotlinx.android.synthetic.main.activity_menu.*
 
 class MenuActivity : AppCompatActivity() {
 
-    var myDB = DatabaseHandler(this)
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_menu)
-        // val token = NetworkUtils.getToken( getIntent().getExtras() )
-
-        val CursorDatabase = myDB.getTokenFromDatabase()
-        var token : String = ""
-
-        if(CursorDatabase.moveToFirst()){
-            token = CursorDatabase.getString(CursorDatabase.getColumnIndex("token"))
-        }
-
-        Log.e("DB", "DADOS-> ${CursorDatabase}")
 
         logoutButton.setOnClickListener {
             val intent = Intent(this, MainActivity::class.java)
-            //intent.putExtra("token", token)
             startActivity(intent)
             finish()
         }
 
         itemMenuColum1Row1.setOnClickListener {
             val intent = Intent(this, ChartListActivity::class.java)
-            //intent.putExtra("token", token)
             startActivity(intent)
             finish()
         }
 
         itemMenuColum1Row2.setOnClickListener {
             val intent = Intent(this, ApprovalsActivity::class.java)
-            //intent.putExtra("token", token)
             startActivity(intent)
             finish()
         }
 
         itemMenuColum2Row1.setOnClickListener {
             val intent = Intent(this, SearchActivity::class.java)
-            //intent.putExtra("token", token)
             startActivity(intent)
             finish()
         }
@@ -62,7 +46,6 @@ class MenuActivity : AppCompatActivity() {
 
         itemMenuColum1Row3.setOnClickListener {
             val intent = Intent(this, WorkCardListActivity::class.java)
-            //intent.putExtra("token", token)
             startActivity(intent)
             finish()
         }
